+++ src/main/java/dev/vitorvidal/characterCreator/domain/model/AbilityScore.java
@@ -1,0 +1,92 @@
+package dev.vitorvidal.characterCreator.domain.model;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+public class AbilityScore {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    private int strength;
+    private int dexterity;
+    private int constitution;
+    private int intelligence;
+    private int wisdom;
+    private int charisma;
+
+    //    @OneToMany(mappedBy = "ability_score")
+    //    private static final List<Character> characters = new ArrayList<>();
+
+    public AbilityScore() {
+    }
+
+    public AbilityScore(int strength, int dexterity, int constitution, int intelligence, int wisdom, int charisma) {
+        this.strength = strength;
+        this.dexterity = dexterity;
+        this.constitution = constitution;
+        this.intelligence = intelligence;
+        this.wisdom = wisdom;
+        this.charisma = charisma;
+    }
+
+    public int getStrength() {
+        return strength;
+    }
+
+    public int getDexterity() {
+        return dexterity;
+    }
+
+    public int getConstitution() {
+        return constitution;
+    }
+
+    public int getIntelligence() {
+        return intelligence;
+    }
+
+    public int getWisdom() {
+        return wisdom;
+    }
+
+    public int getCharisma() {
+        return charisma;
+    }
+
+    public void updateStrength(int strength) {
+        this.strength += strength;
+    }
+
+    public void updateDexterity(int dexterity) {
+        this.dexterity += dexterity;
+    }
+
+    public void updateConstitution(int constitution) {
+        this.constitution += constitution;
+    }
+
+    public void updateIntelligence(int intelligence) {
+        this.intelligence += intelligence;
+    }
+
+    public void updateWisdom(int wisdom) {
+        this.wisdom += wisdom;
+    }
+
+    public void updateCharisma(int charisma) {
+        this.charisma += charisma;
+    }
+
+    public void levelUp() {
+        this.strength += 2;
+        this.dexterity += 2;
+        this.constitution += 2;
+        this.intelligence += 2;
+        this.wisdom += 2;
+        this.charisma += 2;
+    }
+}
